% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeECC}
\alias{computeECC}
\title{computeECC}
\usage{
computeECC(D, maxhomDim, scaleSeq)
}
\description{
computeECC
}
\examples{
N <- 100
set.seed(123)
X <- TDA::circleUnif(N) + rnorm(2*N,mean = 0,sd = 0.2)
D <- TDA::ripsDiag(X,maxdimension = 1,maxscale = 2)$diagram 
scaleSeq = seq(0,2,length.out=11)
computeECC(D,maxhomDim = 1,scaleSeq)
}
