% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computePES}
\alias{computePES}
\title{computePES}
\usage{
computePES(D, homDim, scaleSeq)
}
\description{
computePES
}
\examples{
N <- 100
set.seed(123)
X <- TDA::circleUnif(N) + rnorm(2*N,mean = 0,sd = 0.2)  
D <- TDA::ripsDiag(X,maxdimension = 1,maxscale = 2)$diagram  # compute PD using Rips filtration
scaleSeq = seq(0,2,length.out=11) # sequence of scale values
computePES(D,homDim = 0,scaleSeq) # compute PES for homological dimension H0
computePES(D,homDim = 1,scaleSeq) # compute PES for homological dimension H1
}
